chore: format external url in setting patch
