+++ backend/utils/utils_test.go
@@ -364,3 +364,70 @@ func TestMergeTaskCreateLists(t *testing.T) {
 		})
 	}
 }
+
+func TestNormalizeExternalURL(t *testing.T) {
+	tests := []struct {
+		url     string
+		want    string
+		wantErr bool
+	}{
+		{
+			url:     "http://localhost:3000",
+			want:    "http://localhost:3000",
+			wantErr: false,
+		},
+		{
+			url:     "https://localhost:3000",
+			want:    "https://localhost:3000",
+			wantErr: false,
+		},
+		{
+			url:     "https://localhost",
+			want:    "https://localhost",
+			wantErr: false,
+		},
+		{
+			url:     "http://localhost:80",
+			want:    "http://localhost",
+			wantErr: false,
+		},
+		{
+			url:     "https://localhost:443",
+			want:    "https://localhost",
+			wantErr: false,
+		},
+		{
+			url:     "  https://localhost:3000/ ",
+			want:    "https://localhost:3000",
+			wantErr: false,
+		},
+		// Missing http:// or https://
+		{
+			url:     "localhost:3000",
+			want:    "",
+			wantErr: true,
+		},
+		// Invalid port
+		{
+			url:     "http://localhost:xxx",
+			want:    "",
+			wantErr: true,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.url, func(t *testing.T) {
+			g, err := NormalizeExternalURL(tt.url)
+			if err != nil {
+				if !tt.wantErr {
+					t.Errorf("expect no error, got %s", err.Error())
+				}
+			} else {
+				if tt.wantErr {
+					t.Errorf("expect error")
+				} else if tt.want != g {
+					t.Errorf("expect %s, got %s", tt.want, g)
+				}
+			}
+		})
+	}
+}
